delay loading of the properties for the property condition
There was an issue where, when loading the game server for the first time, the PokemonProperties wouldn't compare properly. I imagine this is because some Cobblemon data didn't have the chance to load before the condition started using PokemonProperties, causing the created property to ignore the missing bits.
Rainbow wooloo apocalypse avoided.